# Reduce the overhead of unconfigured loggers