@swebb2066 2024-02-12
On Windows, the benchmark comparisons using a lock (old) and without locking (new) in AppenderAttachableImpl::appendLoopOnAppenders are:

 | Benchmark | Time | CPU | Time Old | Time New | CPU Old | CPU New | 
 | -------- | ---- | ---- | -------- | -------- | -------- | ---- | 
 | Testing disabled logging request | +0.0034 | +0.1569 | 4 | 4 | 1 | 1 | 
 | Testing disabled logging request/threads:4 | -0.2516 | -0.4159 | 1 | 1 | 4 | 2 | 
 | Logging 5 char string using MessageBuffer, pattern: %m%n | -0.1902 | -0.2120 | 738 | 597 | 443 | 349 | 
 | Logging 5 char string using MessageBuffer, pattern: %m%n/threads:4 | +0.0410 | -0.5887 | 437 | 455 | 1289 | 530 | 
 | Logging 49 char string using MessageBuffer, pattern: %m%n | -0.0180 | -0.1123 | 752 | 739 | 481 | 427 | 
 | Logging 49 char string using MessageBuffer, pattern: %m%n/threads:4 | -0.0485 | -0.3203 | 462 | 439 | 1242 | 844 | 
 | Logging int value using MessageBuffer, pattern: %m%n | -0.0725 | -0.1875 | 1756 | 1629 | 1099 | 893 | 
 | Logging int value using MessageBuffer, pattern: %m%n/threads:4 | -0.1219 | -0.3980 | 740 | 650 | 1744 | 1050 | 
 | Logging int+float using MessageBuffer, pattern: %m%n | -0.1529 | -0.3460 | 3535 | 2995 | 2288 | 1496 | 
 | Logging int+float using MessageBuffer, pattern: %m%n/threads:4 | -0.1841 | -0.3902 | 1387 | 1132 | 3139 | 1914 | 
 | Logging int value using MessageBuffer, pattern: [%d] %m%n | -0.1133 | -0.2317 | 1868 | 1656 | 1144 | 879 | 
 | Logging int value using MessageBuffer, pattern: [%d] [%c] [%p] %m%n | -0.0990 | -0.2480 | 1876 | 1690 | 1308 | 984 | 
 | Logging 49 char string using FMT, pattern: %m%n | -0.0972 | -0.2988 | 796 | 719 | 572 | 401 | 
 | Logging 49 char string using FMT, pattern: %m%n/threads:4 | -0.1717 | -0.4716 | 489 | 405 | 1135 | 600 | 
 | Logging int value using FMT, pattern: %m%n | -0.0587 | -0.0932 | 712 | 670 | 460 | 417 | 
 | Logging int value using FMT, pattern: %m%n/threads:4 | -0.1114 | -0.3957 | 458 | 407 | 1062 | 642 | 
 | Logging int+float using FMT, pattern: %m%n | -0.1281 | -0.1353 | 1132 | 987 | 663 | 573 | 
 | Logging int+float using FMT, pattern: %m%n/threads:4 | -0.1281 | -0.5153 | 544 | 475 | 1662 | 806 | 
 | Async, int value using MessageBuffer, pattern: %m%n | -0.1272 | -0.2650 | 1824 | 1592 | 1224 | 900 | 
 | Async, int value using MessageBuffer, pattern: %m%n/threads:4 | -0.0474 | -0.3363 | 745 | 709 | 1883 | 1250 |