@rm5248 2024-02-12
What's the reasoning behind the lazy initialization of `m_priv`?  It makes the code harder to read IMO since you now have an extra check to see if it is valid, instead of just knowing that it always exists.