@swebb2066 2024-02-12
> What's the reasoning behind the lazy initialization of m_priv? It makes the code harder to read IMO since you now have an extra check to see if it is valid, instead of just knowing that it always exists.

Less memory and processor overhead (see the above benchmarks). 

An application will create many loggers, but few (usually only the root logger) has attached appenders.

 Lazy initialization reduces memory alloc's by avoiding allocating `AppenderAttachableImpl::priv_data` (and initializing a std::mutex + std::vector) for all those loggers that do not have attached appenders.

Throughput is improved because `appendLoopOnAppenders` does not acquire a lock for each decendent logger in the path to the root logger.