@swebb2066 2024-02-12
On Ubuntu (gcc 11), the `benchmark/tools/compare.py --no-color benchmarks /tmp/old.json /tmp/new.json` results are:

| Benchmark | Time | CPU | Time Old | Time New | CPU Old | CPU New | 
 | -------- | ---- | ---- | -------- | -------- | -------- | -------- | 
 | Testing disabled logging request | -0.0358 | -0.0354 | 0 | 0 | 0 | 0 | 
 | Testing disabled logging request/threads:4 | -0.2218 | -0.2217 | 0 | 0 | 1 | 1 | 
 | Logging 5 char string using MessageBuffer, pattern: %m%n | -0.0704 | -0.0703 | 311 | 289 | 311 | 289 | 
 | Logging 5 char string using MessageBuffer, pattern: %m%n/threads:4 | -0.0560 | -0.0531 | 553 | 522 | 2172 | 2056 | 
 | Logging 49 char string using MessageBuffer, pattern: %m%n | -0.0133 | -0.0133 | 318 | 314 | 318 | 314 | 
 | Logging 49 char string using MessageBuffer, pattern: %m%n/threads:4 | -0.2431 | -0.2367 | 575 | 435 | 2262 | 1727 | 
 | Logging int value using MessageBuffer, pattern: %m%n | -0.0474 | -0.0475 | 489 | 466 | 489 | 466 | 
 | Logging int value using MessageBuffer, pattern: %m%n/threads:4 | -0.2029 | -0.2014 | 593 | 473 | 2345 | 1872 | 
 | Logging int+float using MessageBuffer, pattern: %m%n | -0.0000 | -0.0000 | 941 | 941 | 941 | 941 | 
 | Logging int+float using MessageBuffer, pattern: %m%n/threads:4 | -0.2770 | -0.2735 | 754 | 545 | 2997 | 2177 | 
 | Logging int value using MessageBuffer, pattern: [%d] %m%n | -0.1074 | -0.1074 | 627 | 559 | 627 | 559 | 
 | Logging int value using MessageBuffer, pattern: [%d] [%c] [%p] %m%n | -0.1052 | -0.1051 | 629 | 563 | 629 | 563 | 
 | Logging 49 char string using FMT, pattern: %m%n | -0.0333 | -0.0333 | 293 | 283 | 293 | 283 | 
 | Logging 49 char string using FMT, pattern: %m%n/threads:4 | -0.1931 | -0.1891 | 537 | 433 | 2114 | 1714 | 
 | Logging int value using FMT, pattern: %m%n | -0.0156 | -0.0156 | 315 | 310 | 315 | 310 | 
 | Logging int value using FMT, pattern: %m%n/threads:4 | -0.1616 | -0.1590 | 548 | 459 | 2162 | 1818 | 
 | Logging int+float using FMT, pattern: %m%n | -0.0558 | -0.0559 | 468 | 442 | 468 | 442 | 
 | Logging int+float using FMT, pattern: %m%n/threads:4 | +0.0794 | +0.0839 | 429 | 463 | 1700 | 1843 | 
 | Async, int value using MessageBuffer, pattern: %m%n | -0.0463 | -0.0463 | 493 | 471 | 493 | 471 | 
 | Async, int value using MessageBuffer, pattern: %m%n/threads:4 | -0.3025 | -0.2992 | 680 | 474 | 2683 | 1880 |